+++ CHANGELOG.md
@@ -51,6 +51,7 @@ key down... to change the current selected option)
   - we not remove tabIndex=0 on option which are not selected in the group when a click is received inside
   the group => when re-tab to the group, it caused the focus on not last selected option
   - add role radiogroup to the host => we can now add aria-label directly on the host
+- **form-field**: fix accessibility issue with label, description, error message and form control association
 
 ### Security